+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.8.0] - 2023-09-01
+### Added
+- Support for Bulk request with Continue on Error in Detokenize Method
+- Support for Continue on Error in Insert Method
+
 ## [1.7.2] - 2023-08-28
 ### Added
 -  Support for OFF Loglevel.